prototype: mark experimental and isolate backends per the Batch MLA layout (#4697)
- flashinfer/attention/unified.py is now the thin core entry point the
  experimental policy (#4880) asks for: resolve_paged_prefill() and
  UnifiedPagedPrefill.plan()/run() carry @flashinfer_experimental_api
  (first adopter in the repo), value types are lazily re-exported, and
  the implementation is a deferred import.
- flashinfer/experimental/paged_prefill/ holds the implementation, laid
  out like flashinfer/mla/_batch_mla/: _contracts (PlanMetadata,
  Resolution), _planning (validation + derive), _selection (tensor-free
  resolve), _controller (plan/run lifecycle), _backends/ (declarative
  capabilities + fa_backend / cudnn_backend / trtllm_gen_backend).
  "Adapters" are renamed backends; the metadata dict became a frozen
  PlanMetadata.
- plan() publication is transactional: published state swaps only after
  the candidate backend's plan() returned.
- tests move to tests/experimental/ as the policy requires; `attn.backend`
  replaces the private `_backend` the tests and demo poked at.

Verified: tests/experimental on H100 158 passed / 40 skipped / 5 xfailed;
on SM100 153 passed / 45 skipped / 5 xfailed; demo runs on H100.
