+++ PAGED_PREFILL_UNIFICATION_PROPOSAL.md
@@ -2,7 +2,7 @@
 
 **Status**: draft v0.2 (2026-07-14; revised after a three-lens adversarial review: vLLM-maintainer, sglang-maintainer, flashinfer/cuDNN feasibility) · **Scope**: batch prefill with paged KV only — MLA, decode, and ragged prefill are explicit non-goals for v1 (they follow the same pattern later; §6 is explicit about what that scoping means for consumer adoption).
 
-> **Working prototype (2026-07-16)**: branch `yanxu/unified-prefill-prototype` (worktree `.claude/worktrees/unified-prefill-proto`). `flashinfer/attention/unified.py` (contract + adapter layers, `resolve_paged_prefill()`, LSE base-2 contract incl. the cuDNN ln→base-2 fold — first empirical pin of the cuDNN LSE base), a **reject-or-correct fuzzer** (every call either raises a clean error or matches an fp32 oracle; 15 corruption classes, known gaps visible as xfail), conformance matrix green on H100 (fa2/fa3/cudnn/auto: 94 pass) and SM100 Blackwell (fa2/cudnn/trtllm-gen/auto: 93 pass), demo script `prototype_demo_unified_prefill.py`. The prototype's fuzzer found and fixed a live main bug: the cuDNN prefill/decode graph-cache keys omitted `attn_scale` → silent stale-scale replay on same-shape re-plans (fix commit `09c3df82` on the branch, cherry-pickable). Cold-user trial verdict: an engine's attention glue migrates in about a day; blank-file to green in ~30 min.
+> **Working prototype (2026-07-16)**: branch `yanxu/unified-prefill-prototype` (worktree `.claude/worktrees/unified-prefill-proto`). `flashinfer/attention/unified.py` (thin `@flashinfer_experimental_api` entry) + `flashinfer/experimental/paged_prefill/` (contracts / planning / selection / controller / one module per backend, laid out like `flashinfer/mla/_batch_mla/` per #4697; `resolve_paged_prefill()`, LSE base-2 contract — the fuzzer's LSE oracle is what caught the cuDNN base change in #4663), a **reject-or-correct fuzzer** (every call either raises a clean error or matches an fp32 oracle; 15 corruption classes, known gaps visible as xfail), conformance matrix green on H100 (fa2/fa3/cudnn/auto: 94 pass) and SM100 Blackwell (fa2/cudnn/trtllm-gen/auto: 93 pass), demo script `prototype_demo_unified_prefill.py`. The prototype's fuzzer found and fixed a live main bug: the cuDNN prefill/decode graph-cache keys omitted `attn_scale` → silent stale-scale replay on same-shape re-plans (fix commit `09c3df82` on the branch, cherry-pickable). Cold-user trial verdict: an engine's attention glue migrates in about a day; blank-file to green in ~30 min.
 **Companion docs**: `ATTENTION_API_FRAGMENTATION.md` (full 6-family fragmentation survey), PR #3921 (cuDNN token-unit indptr / cu_seq_len direct path).
 
 ---
